+++ NEWS.md
@@ -1,3 +1,9 @@
+# orf 0.1.3
+
+* bug fix for variable importance option
+* bug fix for inference option
+* added unit root tests
+
 # orf 0.1.2
 
 * DESCRIPTION update